// src/app/main/shared/models/Cohorte.ts
export interface Cohorte{
  swCuidate: boolean;
  cuidate: string | null;
  swRcv: boolean;
  rcv: string | null;
  swHipotiroidismo: boolean;
  hipotiroidismo: string | null;
  swArtritisReumatoide: boolean;
  artritisReumatoide: string | null;
  swOncologico: boolean;
  oncologico: string | null;
  swCuidadoColumna: boolean;
  cuidadoColumna: string | null;
  swSiempreFuertes: boolean;
  siempreFuertes: string | null;
  swSaludMental: boolean;
  saludMental: string | null;
}